Atribuição de picking waves

+++ Primavera Integration/SalesOrderPicking/SalesOrderPicking/Lib_Primavera/Model/PickingLine.cs
@@ -8,14 +8,14 @@ namespace SalesOrderPicking.Lib_Primavera.Model {
     public class PickingLine {
 
         private string artigo;
-        private double quantPedida;
+        private int quantPedida;
         private string unidades;
 
         public string Artigo { get { return this.artigo; } }
-        public double QuantPedida { get { return this.quantPedida; } }
+        public int QuantPedida { get { return this.quantPedida; } }
         public string Unidades { get { return this.unidades; } }
 
-        public PickingLine(string artigo, double quantPedida, string unidades) {
+        public PickingLine(string artigo, int quantPedida, string unidades) {
             this.artigo = artigo;
             this.quantPedida = quantPedida;
             this.unidades = unidades;
